# docs/openapi/viron-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Viron API",
    "version": "1.0.0",
    "description": "API specification for the Viron spatial simulation core using DTO-based request/response objects."
  },
  "paths": {
...
    "/api/v1/locations/{id}": {
      "get": {
        "summary": "Get location by ID",
        "parameters": [{ "$ref": "#/components/parameters/LocationIdParam" }],
        "responses": {
          "200": {
            "description": "Location details",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LocationDTO" } }
            }
          },
          "404": {
            "description": "Location not found with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/locations/environment/{environmentId}": {
      "get": {
        "summary": "Get locations in an environment",
        "parameters": [{ "$ref": "#/components/parameters/EnvironmentIdParam" }],
        "responses": {
          "200": {
            "description": "List of locations in the environment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LocationDTO" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locations/grid/{gridId}": {
      "get": {
        "summary": "Get locations in a grid",
        "parameters": [{ "$ref": "#/components/parameters/GridIdParam" }],
        "responses": {
          "200": {
            "description": "List of locations in the grid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LocationDTO" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locations/grid/{gridId}/unoccupied": {
      "get": {
        "summary": "Get unoccupied locations in a grid",
        "parameters": [{ "$ref": "#/components/parameters/GridIdParam" }],
        "responses": {
          "200": {
            "description": "List of unoccupied locations in the grid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LocationDTO" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/locations/{locationId}/neighbors": {
      "get": {
        "summary": "Get the locations adjacent to a location within the same grid",
        "parameters": [{ "$ref": "#/components/parameters/LocationIdParam" }],
        "responses": {
          "200": {
            "description": "List of adjacent locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LocationDTO" }
                }
              }
            }
          },
          "404": {
            "description": "Location not found, or not in any grid",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/locations/{locationId}/entity/{entityId}": {
      "put": {
        "summary": "Add entity to location",
        "description": "Places an unplaced entity at the location. An entity occupies at most one location, so the request is a no-op when the entity is already at that location and a conflict when it is placed elsewhere.",
        "parameters": [
          { "$ref": "#/components/parameters/LocationIdParam" },
          { "$ref": "#/components/parameters/EntityIdParam" }
        ],
        "responses": {
          "200": { "description": "Entity added, or already present at that location" },
          "404": {
            "description": "Location or entity not found with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "409": {
            "description": "Entity is already placed at a different location",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove entity from location",
        "parameters": [
          { "$ref": "#/components/parameters/LocationIdParam" },
          { "$ref": "#/components/parameters/EntityIdParam" }
        ],
        "responses": {
          "204": { "description": "Entity removed" },
          "404": {
            "description": "Location not found with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/locations/{locationId}/entity/{entityId}/move": {
      "put": {
        "summary": "Move an entity to an adjacent, unoccupied location in the same grid",
        "parameters": [
          { "$ref": "#/components/parameters/LocationIdParam" },
          { "$ref": "#/components/parameters/EntityIdParam" }
        ],
        "responses": {
          "204": { "description": "Entity moved" },
          "400": {
            "description": "Target location is not in the same grid as the entity, or is not adjacent to the entity's current location",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": {
            "description": "Entity is not placed at any location, or target location not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "409": {
            "description": "Target location is already occupied",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/locations/{locationId}/entities": {
      "get": {
        "summary": "Get entity IDs at a location",
        "parameters": [{ "$ref": "#/components/parameters/LocationIdParam" }],
        "responses": {
          "200": {
            "description": "List of entity IDs at the location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "integer" }
                }
              }
            }
          },
          "404": {
            "description": "Location not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/locations/{locationId}/occupied": {
      "get": {
        "summary": "Check whether a location is occupied",
        "parameters": [{ "$ref": "#/components/parameters/LocationIdParam" }],
        "responses": {
          "200": {
            "description": "Whether the location is occupied",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "404": {
            "description": "Location not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/api/v1/locations/entity/{entityId}": {
      "get": {
        "summary": "Get the location of an entity",
        "parameters": [{ "$ref": "#/components/parameters/EntityIdParam" }],
        "responses": {
          "200": {
            "description": "Location of the entity",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LocationDTO" } }
            }
          },
          "404": {
            "description": "Location not found for entity",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove entity from current location",
        "parameters": [{ "$ref": "#/components/parameters/EntityIdParam" }],
        "responses": {
          "204": { "description": "Entity removed" },
          "404": {
            "description": "Location not found for entity",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
...
    "/api/v1/entities/location/{locationId}": {
      "get": {
        "summary": "Get entities in a location",
        "parameters": [{ "$ref": "#/components/parameters/LocationIdParam" }],
        "responses": {
          "200": {
            "description": "List of entities in the location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EntityDTO" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LocationDTO": {
        "type": "object",
        "properties": {
          "locationId": { "type": "integer" },
          "x": { "type": "integer" },
          "y": { "type": "integer" }
        }
      },
      "EntityDTO": {
        "type": "object",
        "properties": {
          "entityId": { "type": "integer" },
          "name": { "type": "string" },
          "creationDate": { "type": "string" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response body, returned for every 4xx and 5xx response.",
        "properties": {
          "status": { "type": "integer", "description": "HTTP status code" },
          "message": { "type": "string", "description": "Error message describing what went wrong" }
        }
      }
    },
    "parameters": {
      "EnvironmentIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "GridIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "LocationIdParam": {
        "name": "locationId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "EntityIdParam": {
        "name": "entityId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
...
    }
  }
}